requirements.txt
aiohttp==3.10.6
SQLAlchemy==2.0.37
python-dotenv==1.0.1
## The following requirements were added by pip freeze:
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiomysql==0.2.0
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncio==3.4.3
asynciolimiter==1.0.0
attrs==24.2.0
Authlib==1.3.2
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
croniter==3.0.3
cryptography==43.0.1
dataclasses==0.6
Deprecated==1.2.14
discord-py-interactions==5.14.0
discord-typings==0.9.0
distro==1.9.0
emoji==2.12.1
ERAlchemy==1.2.10
exceptiongroup==1.2.2
Flask==3.0.3
frozenlist==1.4.1
git-filter-repo==2.45.0
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.65.0
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.8
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.10.0
Mako==1.3.5
Markdown==3.7
MarkupSafe==2.1.5
more-itertools==9.1.0
msgspec==0.18.6
multidict==6.0.5
oauthlib==3.2.2
osrsreboxed==2.3.30
pillow==10.4.0
priority==2.0.0
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.10
py-spy==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic<=2.0.9
pydantic_core==2.33.2
PyGithub==2.4.0
PyJWT==2.9.0
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
Quart==0.19.6
quart-cors==0.8.0
Quart-JWT-Extended==0.1.0
quart-rate-limiter==0.11.0
ratelimiter==1.2.0.post0
redis==5.0.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
setuptools==60.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlacodegen-v2==0.1.4
tabulate==0.9.0
taskgroup==0.0.0a4
tomli==2.0.1
tqdm==4.67.1
typeguard==4.4.1
typing-inspection==0.4.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
websockets==13.1
Werkzeug==3.0.4
wom.py==1.0.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.12.1
